requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langgraph>=0.0.20
openai>=1.0.0
azure-search-documents>=11.5.0
azure-core>=1.29.0
python-dotenv>=1.0.0
pandas>=2.0.0
pyodbc>=4.0.39
psycopg2-binary>=2.9.0
